#2145 Move video recorder JUnit/TestNG implementations to just "selenide-video-recorder"
we don't want our users to bother with too many dependencies.

But integration tests for JUnit/TestNG are still kept in separate modules (at least for now).
