--- modules/video-recorder-core/build.gradle
@@ -1,24 +1,0 @@
-plugins {
-  id 'java'
-}
-
-ext {
-  artifactId = 'video-recorder-core'
-}
-
-repositories {
-  mavenCentral()
-}
-
-dependencies {
-  api project(":statics")
-  testImplementation project(':statics').sourceSets.test.output
-  testImplementation project(':modules:core').sourceSets.test.output
-  testImplementation("org.junit.jupiter:junit-jupiter-api:$junitVersion")
-  testImplementation("org.junit.jupiter:junit-jupiter-engine:$junitVersion")
-  testImplementation("org.assertj:assertj-core:$assertjVersion") { transitive false }
-  implementation("org.bytedeco:javacv:$javacvVersion")
-  implementation("org.bytedeco:javacv-platform:$javacvVersion")
-}
-
-apply from: rootProject.file('gradle/publish-module.gradle')
