+++ modules/video-recorder/src/main/java/org/selenide/videorecorder/junit5/RecordVideoJunit.java
@@ -1,0 +1,18 @@
+package org.selenide.videorecorder.junit5;
+
+import org.junit.jupiter.api.extension.ExtendWith;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Created by Serhii Bryt
+ * 30.04.2024 12:56
+ **/
+@ExtendWith({BrowserRecorderCallBack.class})
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface RecordVideoJunit {
+}
