+++ settings.gradle
@@ -10,6 +10,6 @@ include ':modules:testng'
 include ':modules:clear-with-shortcut'
 include ':modules:full-screenshot'
 include ':modules:error-message-customizer'
-include ':modules:video-recorder-core'
+include ':modules:video-recorder'
 include ':modules:video-recorder-junit'
 include ':modules:video-recorder-testng'
